Remove duplicate copy of Spyc.php
