+++ public_html/events.php
@@ -4,7 +4,7 @@ $subtitle = 'Details of past and future nf-core meetups.';
 $md_github_url = 'https://github.com/nf-core/nf-co.re/blob/master/nf-core-events.yaml';
 include('../includes/header.php');
 
-require_once("../Spyc.php");
+require_once("../includes/libraries/Spyc.php");
 $events_yaml = spyc_load_file('../nf-core-events.yaml');
 $events = $events_yaml['events'];
